ab_contracts_test_utils::dummy_wallet::tx_handler_ffi::authorize

Struct InternalArgs

#[repr(C)]
pub struct InternalArgs<'internal_args> {
    pub _env_ptr: NonNull<Env<'internal_args>>,
    pub _header_ptr: NonNull<<TransactionHeader as IoType>::PointerType>,
    pub _header_size: NonNull<u32>,
    pub _read_slots_ptr: NonNull<<TxHandlerSlots as IoType>::PointerType>,
    pub _read_slots_size: NonNull<u32>,
    pub _write_slots_ptr: NonNull<<TxHandlerSlots as IoType>::PointerType>,
    pub _write_slots_size: NonNull<u32>,
    pub _payload_ptr: NonNull<<TxHandlerPayload as IoType>::PointerType>,
    pub _payload_size: NonNull<u32>,
    pub _seal_ptr: NonNull<<TxHandlerSeal as IoType>::PointerType>,
    pub _seal_size: NonNull<u32>,
    /* private fields */
}

Data structure containing expected input to tx_handler_authorize(), it is used internally by the contract, there should be no need to construct it explicitly except maybe in contract’s own tests

Fields
_env_ptr: NonNull<Env<'internal_args>>_header_ptr: NonNull<<TransactionHeader as IoType>::PointerType>_header_size: NonNull<u32>

Size of the contents _header_ptr points to

_read_slots_ptr: NonNull<<TxHandlerSlots as IoType>::PointerType>_read_slots_size: NonNull<u32>

Size of the contents _read_slots_ptr points to

_write_slots_ptr: NonNull<<TxHandlerSlots as IoType>::PointerType>_write_slots_size: NonNull<u32>

Size of the contents _write_slots_ptr points to

_payload_ptr: NonNull<<TxHandlerPayload as IoType>::PointerType>_payload_size: NonNull<u32>

Size of the contents _payload_ptr points to

_seal_ptr: NonNull<<TxHandlerSeal as IoType>::PointerType>_seal_size: NonNull<u32>

Size of the contents _seal_ptr points to